# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeerSetu API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  user: "user",
  provider: "provider",
  ngo: "ngo",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: RegisterRequestRole;
  location?: string;
  phone?: string;
  serviceArea?: string;
  organizationName?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  provider: "provider",
  ngo: "ngo",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  location?: string;
  phone?: string;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Provider {
  id: number;
  userId?: number;
  name: string;
  email: string;
  phone?: string;
  serviceArea: string;
  city?: string;
  state?: string;
  portfolio?: string;
  experience?: number;
  rating?: number;
  reviewCount?: number;
  ngoId?: number | null;
  createdAt?: string;
}

export interface ProviderUpdateRequest {
  phone?: string;
  serviceArea?: string;
  city?: string;
  state?: string;
  portfolio?: string;
  experience?: number;
}

export type ServiceRequestStatus =
  (typeof ServiceRequestStatus)[keyof typeof ServiceRequestStatus];

export const ServiceRequestStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
} as const;

export interface ServiceRequest {
  id: number;
  userId: number;
  providerId?: number | null;
  requestType: string;
  systemType: string;
  location: string;
  preferredDate?: string;
  status: ServiceRequestStatus;
  notes?: string;
  rating?: number | null;
  review?: string | null;
  userName?: string;
  providerName?: string | null;
  createdAt?: string;
}

export interface CreateServiceRequestRequest {
  providerId?: number | null;
  requestType: string;
  systemType: string;
  location: string;
  preferredDate?: string;
  notes?: string;
}

export type UpdateStatusRequestStatus =
  (typeof UpdateStatusRequestStatus)[keyof typeof UpdateStatusRequestStatus];

export const UpdateStatusRequestStatus = {
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
}

export interface RatingRequest {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  review?: string;
}

export interface SaveProviderRequest {
  providerId: number;
}

export interface Ngo {
  id: number;
  userId?: number;
  name: string;
  email: string;
  organizationName?: string;
  providerCount?: number;
  totalInstallations?: number;
  createdAt?: string;
}

export type CampaignStatus =
  (typeof CampaignStatus)[keyof typeof CampaignStatus];

export const CampaignStatus = {
  active: "active",
  completed: "completed",
  planned: "planned",
} as const;

export interface Campaign {
  id: number;
  ngoId?: number;
  title: string;
  description: string;
  targetAudience?: string;
  startDate?: string;
  endDate?: string;
  status?: CampaignStatus;
  createdAt?: string;
}

export interface CreateCampaignRequest {
  title: string;
  description: string;
  targetAudience?: string;
  startDate?: string;
  endDate?: string;
}

export interface InviteProviderRequest {
  email: string;
  name: string;
  phone?: string;
}

export interface GovernmentScheme {
  id: number;
  name: string;
  state: string;
  subsidy: string;
  eligibility?: string;
  description?: string;
  link?: string;
  createdAt?: string;
}

export interface Article {
  id: number;
  title: string;
  category: string;
  content: string;
  imageUrl?: string;
  tags?: string[];
  createdAt?: string;
}

export type GetProvidersParams = {
  city?: string;
  state?: string;
};

export type GetSchemesParams = {
  state?: string;
};

export type GetArticlesParams = {
  category?: string;
};