# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NeerSetu API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /providers:
    get:
      operationId: getProviders
      tags: [providers]
      summary: List all providers
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
  /providers/me:
    get:
      operationId: getMyProviderProfile
      tags: [providers]
      summary: Get my provider profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Provider profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
    put:
      operationId: updateProviderProfile
      tags: [providers]
      summary: Update my provider profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderUpdateRequest"
      responses:
        "200":
          description: Updated provider profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
  /providers/{id}:
    get:
      operationId: getProviderById
      tags: [providers]
      summary: Get provider by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Provider details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
...
  /ngos/me:
    get:
      operationId: getNgoProfile
      tags: [ngos]
      summary: Get my NGO profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: NGO profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ngo"
  /ngos/me/providers:
    get:
      operationId: getNgoProviders
      tags: [ngos]
      summary: Get providers onboarded by NGO
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Providers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
  /ngos/me/campaigns:
    get:
      operationId: getNgoCampaigns
      tags: [ngos]
      summary: Get NGO campaigns
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Campaigns list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [ngos]
      summary: Create awareness campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "200":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /ngos/invite-provider:
    post:
      operationId: inviteProvider
      tags: [ngos]
      summary: Invite a provider
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteProviderRequest"
      responses:
        "200":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /schemes:
    get:
      operationId: getSchemes
      tags: [schemes]
      summary: Get government schemes
      parameters:
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of schemes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GovernmentScheme"
  /articles:
    get:
      operationId: getArticles
      tags: [articles]
      summary: Get awareness articles
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [user, provider, ngo]
        location:
          type: string
        phone:
          type: string
        serviceArea:
          type: string
        organizationName:
          type: string
      required:
        - name
        - email
        - password
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, provider, ngo]
        location:
          type: string
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
    Provider:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        serviceArea:
          type: string
        city:
          type: string
        state:
          type: string
        portfolio:
          type: string
        experience:
          type: integer
        rating:
          type: number
          format: float
        reviewCount:
          type: integer
        ngoId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - serviceArea
    ProviderUpdateRequest:
      type: object
      properties:
        phone:
          type: string
        serviceArea:
          type: string
        city:
          type: string
        state:
          type: string
        portfolio:
          type: string
        experience:
          type: integer
...
    Ngo:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        organizationName:
          type: string
        providerCount:
          type: integer
        totalInstallations:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
    Campaign:
      type: object
      properties:
        id:
          type: integer
        ngoId:
          type: integer
        title:
          type: string
        description:
          type: string
        targetAudience:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, completed, planned]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
    CreateCampaignRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        targetAudience:
          type: string
        startDate:
          type: string
        endDate:
          type: string
      required:
        - title
        - description
    InviteProviderRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
      required:
        - email
        - name
    GovernmentScheme:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
        subsidy:
          type: string
        eligibility:
          type: string
        description:
          type: string
        link:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - state
        - subsidy
    Article:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - content